requirements.txt
click==8.1.7
emoji==2.10.0
hydra-core==1.3.2
keybert==0.8.3
matplotlib==3.8.2
nltk==3.8.1
numpy==1.24.1
omegaconf==2.3.0
pandas==2.2.0
pycountry==23.12.11
scipy==1.12.0
seaborn==0.13.1
setuptools==68.2.2
torch==2.1.2+cu118
transformers==4.37.0
